incorporate freebsd valgrind port from https://github.com/kgudka/valgrind-freebsd-git

+++ valgrind-3.7.0/memcheck/tests/x86/more_x86_fp.c
@@ -87,7 +87,7 @@ void test_fcvt(double a)
     long double la;
     int16_t fpuc;
     int i;
-    int64_t lla;
+    long long int lla;
     int ia;
     int16_t wa;
     double ra;
